# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记忆 API",
        "title": "Remembrance API",
        "contact": {
            "name": "lzd",
            "email": "2804366305@qq.com"
        },
        "version": "1.0"
    },
    "host": "8.138.81.141",
    "basePath": "api",
    "paths": {
        "/api/check_code": {
            "post": {
                "description": "根据发送的 邮箱 请求方式（register，change）检查 验证码， 会返回 \"验证码不存在或已过期\"/\"验证码正确\"/\"验证码错误\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "检查验证码",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/email.Message"
                        }
                    },
                    {
                        "description": "code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/email.Message"
                        }
                    },
                    {
                        "description": "请求类型:注册'register',改密码'change'",
                        "name": "gettype",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/email.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"成功\"}",
                        "schema": {
                            "$ref": "#/definitions/controller.okmesdata"
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"Failure\"}",
                        "schema": {
                            "$ref": "#/definitions/controller.failmesdata"
                        }
                    }
                }
            }
        },
...
        "/api/get_code": {
            "get": {
                "description": "根据情况向指定邮箱发送不同时限的验证码，",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "获取验证码",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/email.Message"
                        }
                    },
                    {
                        "description": "请求类型:注册'register',改密码'change'",
                        "name": "gettype",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/email.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/controller.okmesdata"
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"Failure\"}",
                        "schema": {
                            "$ref": "#/definitions/controller.failmesdata"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "获取email password 并检查是否匹配， 匹配成功后会将成功消息与 userid,token(HMAC SHA256 算法创建,密钥为\"remembrance\") 传回前端, userid需要解析才能获得， 在之后会用到",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.okmesdata"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.failmesdata"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "put": {
                "description": "前端检查验证码正确之后，再将 email 与 password 传给后端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.okmesdata"
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"Failure\"}",
                        "schema": {
                            "$ref": "#/definitions/controller.failmesdata"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.failmesdata": {
            "type": "object"
        },
        "controller.okmesdata": {
            "type": "object"
        },
        "email.Message": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "gettype": {
                    "description": "请求类型 注册\"register\" 改密码\"change\"",
                    "type": "string"
                }
            }
        },
...
        "models.S_User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "pospersonaltnum": {
                    "type": "integer"
                },
                "postcommonnum": {
                    "type": "integer"
                },
                "postgroupnum": {
                    "type": "integer"
                },
                "stampnum": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "DeletedAt DeletedAt `gorm:\"index\"`",
                    "type": "string"
                }
            }
        },
...
    }
}